Rename reports following call w Richard
