+++ src/lib/queries/uncreditedInactiveLicences.js
@@ -1,11 +1,16 @@
+const moment = require('moment');
+
 module.exports = `with tempTable as (select
-distinct l.licence_ref as "Licence No.",
+distinct 
+    null as "Inactive licences not yet credited. Data generated ${moment().format('DD/MM/YYYY')}",
+    l.licence_ref as "Licence No.",
     cv.charge_version_id as "Charge Version ID",
     ce.charge_element_id as "Charge Element ID",
     ce.description as "Charge Element Description",
     companies.name as "Licence Holder Name",
     l.start_date as "Orig. Effective Date",
     l.expired_date as "Expiry date",
+    l.revoked_date as "Revoked date",
     ce.time_limited_start_date as "TL Start",
     ce.time_limited_end_date as "TL End",
     cv.invoice_account_id as "Invoice Account ID",
@@ -49,16 +54,19 @@ or v2doc.end_date >= NOW())
 join crm_v2.companies companies on
 companies.company_id = v2docRoles.company_id
 where
-(l.expired_date is null
-or l.expired_date >= NOW())
-and (l.lapsed_date is null
-or l.lapsed_date >= NOW())
-and (l.revoked_date is null
-or l.revoked_date >= NOW())
+(l.revoked_date < NOW() OR l.lapsed_date < NOW() or l.expired_date < NOW())
 and (cv.end_date is null
-or cv.end_date >= NOW())) select * from tempTable where "Charge Version ID" not in 
-(select charge_version_id from water.billing_batch_charge_version_years bbcvy 
-join water.billing_batches bb on 
-bb.billing_batch_id = bbcvy.billing_batch_id  where 
-bbcvy.financial_year_ending = $1 
-and bb.status = 'sent')`;
+or cv.end_date >= NOW())) select * from tempTable where "Charge Version ID" in
+(select charge_version_id from water.billing_batch_charge_version_years bbcvy
+join water.billing_batches bb on
+bb.billing_batch_id = bbcvy.billing_batch_id  where
+bbcvy.financial_year_ending = $1
+and bb.status = 'sent'
+and bb.batch_type= 'annual')
+and "Charge Version ID" not in (
+    select charge_version_id from water.billing_batch_charge_version_years bbcvy
+join water.billing_batches bb on
+bb.billing_batch_id = bbcvy.billing_batch_id  where
+bbcvy.financial_year_ending = $1
+and bb.status = 'sent'
+and bb.batch_type= 'supplementary')`;
