+++ src/lib/reports-handlers/uncreditedInactiveLicencesReportHandler.js
@@ -1,5 +1,5 @@
 const { pool } = require('../connectors/db');
-const revokedUncreditedLicences = require('../queries/revokedUncreditedLicences');
+const uncreditedInactiveLicences = require('../queries/uncreditedInactiveLicences');
 const { logger } = require('../../logger');
 const csvWritingHelper = require('../helpers/write-csv-to-s3');
 const { getFinancialYear } = require('../helpers/financial-year');
@@ -7,9 +7,9 @@ const handler = async () => {
   try {
     const financialYear = getFinancialYear(new Date());
     // Run query
-    const res = await pool.query(revokedUncreditedLicences, [financialYear]);
+    const res = await pool.query(uncreditedInactiveLicences, [financialYear]);
     // Write results to S3
-    const filename = 'revokedUncreditedLicences.csv';
+    const filename = 'uncreditedInactiveLicences.csv';
     await csvWritingHelper.generateCsv(filename, res.rows, res.fields);
   } catch (e) {
     logger.error(e);
